Smell Density implementation and some Refactoring
Code Refactoring:
- EagerTestStrategy : added getActionForStats() in order to create an Action and store it in a Session.
- GeneralFixtureStrategy: added getActionForStats() in order to create an Action and store it in a Session.
- LackOfCohesionStrategy : added getActionForStats() in order to create an Action and store it in a Session.

Smell Density:
 Class Session.java:
- Adding attributes "nOfTotalMethod" because in Eager Test we have to calculate the Density of smell and Eager Test works with method.

-Added following methods in order to calculate the density of every smell: densityET(), densityLOC(), densityGF()

Class IDetector: added the method getMethodBeansNumber() and it was implemented in "TestSmellStructuralDetector" and "TestSmellTextualDetector"

Class TestSmellStructuralDetector: added the logic in which for every test class, we need to know the method inside and and in the end count it (with the method getMethodBeansNumber ). The same thing for TestSmellTextualDetector.

class StatsPanel: creation of the panel for the UI visualization.

class ConverterUtilities: change the getMethodFromClass() method's visibility to public in order to call this method inside the class TestSmellStructuralDetector and TestSmellTextualDetector in order to obtain the method of every test class and finally count it.